requirements.txt
djangorestframework
markdown
django-filter
requests